Login popup added
The login popup was added to the LoginPage, right now it  consists of a form and a Login component outside the form that is the one responsible of the popup window, due to some problems with the popup it could not be added to the form nor the LoginButton component. Further investigations will be done.