Added user's name retrieval
Added the class UserData in the package data-access inside source, that has the method GetUserName().
This method uses the hook "useLDflexValue" in order to retrieve the name of the user authenticated.
In order to test this functionality, it has been added to Login.js a paragraph from which the function is called "<p>{GetUserName()}</p>"