Added a notification system
Actually, not sure whether this is the final version. Right now, using the useNotification hook, a notification is being created, with .ttl format.
I have several problems here, despite the notification is actually created, a null pointer exception is triggered. Also, with the idea of following the viadeSpec, the notifications should be .jsonld